text field updates
