+++ frontend/styles/organisms.NewMeeting.module.scss
@@ -1,0 +1,73 @@
+.calendarNavbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px 20px;
+    background-color: #f8f9fa;
+    border-bottom: 1px solid #dee2e6;
+  }
+  
+  .dateContainer {
+    display: flex;
+    align-items: center;
+  
+    .navButton {
+      background-color: transparent;
+      border: none;
+      font-size: 18px;
+      cursor: pointer;
+      margin: 0 10px;
+    }
+  
+    .currentDate {
+      font-weight: bold;
+      font-size: 18px;
+    }
+  }
+  
+  .viewButtons {
+    display: flex;
+  
+    .viewButton {
+      background-color: transparent;
+      border: none;
+      cursor: pointer;
+      padding: 5px 10px;
+      font-size: 16px;
+      margin: 0 5px;
+      border-radius: 4px;
+      transition: background-color 0.2s ease-in-out;
+  
+      &:hover {
+        background-color: #e9ecef;
+      }
+    }
+  }
+  
+  .actionButtons {
+    display: flex;
+    align-items: center;
+  
+    .todayButton {
+      background-color: #007bff;
+      color: white;
+      border: none;
+      border-radius: 4px;
+      padding: 5px 10px;
+      margin-right: 10px;
+      cursor: pointer;
+      transition: background-color 0.2s ease-in-out;
+  
+      &:hover {
+        background-color: #0056b3;
+      }
+    }
+  
+    .searchButton {
+      background-color: transparent;
+      border: none;
+      font-size: 20px;
+      cursor: pointer;
+    }
+  }
+
